added float to gold

+++ app/src/main/java/com/mgApp/FragmentAddRakam.kt
@@ -70,7 +70,7 @@ class FragmentAddRakam : Fragment() {
             if (rakamType.isEmpty() or rakamWeight.isEmpty() or rakamNetWeight.isEmpty()){
                 Toast.makeText(activity,"Please insert Valid/Complete Data!", Toast.LENGTH_LONG).show()
             }
-            else if (rakamWeight.toInt() > rakamNetWeight.toInt()){
+            else if (rakamWeight.toFloat() > rakamNetWeight.toFloat()){
                 Toast.makeText(activity,"Fine weight should be less than or equal to net weight!", Toast.LENGTH_LONG).show()
             }
             else{
